Troub 69 - Internationalisation (#41)
* mise en place des fichiers de langue

* changement affichage

* ajout traductions

* fix image

* affichage et langues

* Fix btn deconnexion

* margin pour le form

* uppercase

* supression de l'image quand on modifie ou supprime un lieu

* ajout d'un fichier pour la liste des langues de notre système

* utilisation foreach pour parcourir les langues définies dans notre fichier de config

* nom accueil pour lieux

* accueil en lieux

* ajout de clés

* v.rification user pour modifier et supprimer

* création des clés et ajout dans les pages

* fichier de langue validations

* traduction pour supression

* suppression remove au lieu de reload et messages erreur fichiers de langues

* npm run format

* image par defaut dans le storage public

---------

+++ app/Http/Controllers/LieuxController.php
@@ -12,10 +12,9 @@ use App\Models\Ville;
 use Exception;
 use Illuminate\Database\QueryException;
 use App\Http\Requests\LieuRequest;
-use App\Http\Requests\LieuModifierRequest;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Str;
+use Illuminate\Support\Facades\Storage;
 
 
 class LieuxController extends Controller
@@ -85,46 +84,44 @@ class LieuxController extends Controller
     }
 
     public function AjouterUnLieu(LieuRequest $request)
-    {
-        try {
-            $lieu = new Lieu();
-            $lieu->rue = $request->rue;
-            $lieu->noCivic = $request->noCivic;
-            $lieu->codePostal = (strtoupper($request->codePostal));
-            $lieu->nomEtablissement = $request->nomEtablissement;
-            //TODO Trouver comment stocker les images
-            if ($request->hasFile('photoLieu')) {
-                $file = $request->file('photoLieu');
+{
+    try {
+        $lieu = new Lieu();
+        $lieu->rue = $request->rue;
+        $lieu->noCivic = $request->noCivic;
+        $lieu->codePostal = (strtoupper($request->codePostal));
+        $lieu->nomEtablissement = $request->nomEtablissement;
 
-                if ($file->isValid()) {
-                    $originalName = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-                    $extension = $file->getClientOriginalExtension();
-                    $fileName = time() . '_' . Str::slug($originalName) . '.' . $extension;
-
-                    $file->move(public_path('Images/lieux'), $fileName);
-
-                    $lieu->photoLieu = 'Images/lieux/' . $fileName;
-                }
-            } else {
-                $lieu->photoLieu = 'Images/lieux/image_defaut.png';
-            }
-
-            $lieu->siteWeb = $request->siteWeb;
-            $lieu->numeroTelephone = $request->numeroTelephone;
-            $lieu->actif = true;
-            $lieu->description = $request->description;
-            $lieu->quartier_id = $request->selectQuartierLieu;
-            $lieu->typeLieu_id = $request->selectTypeLieu;
-            $lieu->proprietaire_id = Auth::id();
-            $lieu->save();
-
-            session()->flash('formulaireValide', 'true');
-            return redirect()->route('usagerLieux.afficher');
-        } catch (\Exception $e) {
-            Log::error("Erreur lors de l'ajout d'un lieu: " . $e->getMessage());
-            return redirect()->route('usagerLieux.afficher');
+        $photoDefautPath = 'lieux/image_defaut.png';
+        if (!Storage::disk('DevActivite')->exists($photoDefautPath)) {
+            Storage::disk('DevActivite')->put($photoDefautPath, file_get_contents(public_path('Images/lieux/image_defaut.png')));
         }
+
+        if ($request->hasFile('photoLieu')) {
+            $file = $request->file('photoLieu');
+            $chemin = $file->store('lieux', 'DevActivite');
+            $lieu->photoLieu = $chemin;
+        } else {
+            $lieu->photoLieu = $photoDefautPath;
+        }
+
+        $lieu->siteWeb = $request->siteWeb;
+        $lieu->numeroTelephone = $request->numeroTelephone;
+        $lieu->actif = true;
+        $lieu->description = $request->description;
+        $lieu->quartier_id = $request->selectQuartierLieu;
+        $lieu->typeLieu_id = $request->selectTypeLieu;
+        $lieu->proprietaire_id = Auth::id();
+        $lieu->save();
+
+        session()->flash('formulaireValide', 'true');
+        return redirect()->route('usagerLieux.afficher');
+    } catch (\Exception $e) {
+        Log::error("Erreur lors de l'ajout d'un lieu: " . $e->getMessage());
+        return redirect()->route('usagerLieux.afficher');
     }
+}
+
 
     /**
      * Display the specified resource.
@@ -137,7 +134,6 @@ class LieuxController extends Controller
         $activites = Activite::whereIn("id", $idActivites)->where('actif', 1)->get();
 
         return view('zoomLieu', compact('lieuActuel', 'activites'));
-
     }
 
     public function ObtenirUnLieu(Request $request)
@@ -160,53 +156,74 @@ class LieuxController extends Controller
     public function ModifierUnLieu(LieuRequest $request, string $id)
     {
         $lieu = Lieu::findOrFail($id);
+        $utilisateur = auth()->user();
+        $estAdmin = $utilisateur->role->nom === 'admin';
+        $estProprietaire = $lieu->proprietaire_id === $utilisateur->id;
+        if (!$estProprietaire && !$estAdmin) {
+            return redirect()->route('usagerLieux.afficher');
+        }
+    
         try {
+            $photoDefautPath = 'lieux/image_defaut.png';
+            if (!Storage::disk('DevActivite')->exists($photoDefautPath)) {
+                Storage::disk('DevActivite')->put($photoDefautPath, file_get_contents(public_path('Images/lieux/image_defaut.png')));
+            }
+    
             $lieu->rue = $request->rue;
             $lieu->noCivic = $request->noCivic;
             $lieu->codePostal = $request->codePostal;
             $lieu->nomEtablissement = $request->nomEtablissement;
-            //TODO Trouver comment stocker les images
-            if ($request->hasFile('photoLieu')) {
-                $file = $request->file('photoLieu');
-
-                if ($file->isValid()) {
-                    if ($lieu->photoLieu && $lieu->photoLieu !== 'Images/lieux/image_defaut.png' && file_exists(public_path($lieu->photoLieu))) {
-                        unlink(public_path($lieu->photoLieu));
-                    }
-
-                    $originalName = pathinfo($file->getClientOriginalName(), PATHINFO_FILENAME);
-                    $extension = $file->getClientOriginalExtension();
-                    $fileName = time() . '_' . Str::slug($originalName) . '.' . $extension;
-
-                    $file->move(public_path('Images/lieux'), $fileName);
-
-                    $lieu->photoLieu = 'Images/lieux/' . $fileName;
-                }
-            }
             $lieu->siteWeb = $request->siteWeb;
             $lieu->numeroTelephone = $request->numeroTelephone;
             $lieu->description = $request->description;
             $lieu->quartier_id = $request->selectQuartierLieu;
             $lieu->typeLieu_id = $request->selectTypeLieu;
+    
+            if ($request->hasFile('photoLieu')) {
+                if ($lieu->photoLieu && $lieu->photoLieu !== $photoDefautPath) {
+                    Storage::disk('DevActivite')->delete($lieu->photoLieu);
+                }
+    
+                $file = $request->file('photoLieu');
+                $chemin = $file->store('lieux', 'DevActivite');
+                $lieu->photoLieu = $chemin;
+            }
+    
+            if (!$request->hasFile('photoLieu') && $lieu->photoLieu === $photoDefautPath) {
+                $lieu->photoLieu = $photoDefautPath;
+            }
+    
+            // Sauvegarde des modifications
             $lieu->save();
-
+    
             session()->flash('formulaireModifierValide', 'true');
             return redirect()->route('usagerLieux.afficher');
         } catch (\Exception $e) {
             Log::error("Erreur lors de la modification d'un lieu: " . $e->getMessage());
-            return redirect()->route('usagerLieux.afficher');
+            return redirect()->route('usagerLieux.afficher')->with('error',  __('erreurGenerale'));
         }
     }
+    
 
     public function SupprimerUnLieu($id)
-    {
+    { 
+        $lieu = Lieu::findOrFail($id);
+        $utilisateur = auth()->user();
+        $estAdmin = $utilisateur->role->nom === 'admin';
+        $estProprietaire = $lieu->proprietaire_id === $utilisateur->id;
+        if (!$estProprietaire && !$estAdmin) {
+            return response()->json(['success' => false, 'message' => 'Accès refusé.'], 403);
+        }
         try {
-            $lieu = Lieu::findOrFail($id);
+           
+            if ($lieu->photoLieu && $lieu->photoLieu !== 'Images/lieux/image_defaut.png') {
+                Storage::disk('DevActivite')->delete($lieu->photoLieu);
+            }
             $lieu->delete();
-            return response()->json(["success" => true, "message" => "Lieu supprimé avec succès."]);
+            return response()->json(["success" => true, "message" => __('succesSupprimer')]);
         } catch (\Exception $e) {
             Log::error("Erreur lors de la suppression d'un lieu: " . $e->getMessage());
-            return response()->json(["success" => false, "message" => "Erreur lors de la suppression."], 500);
+            return response()->json(["success" => false, "message" =>  __('erreurSuppresion')], 500);
         }
     }
 }
